修正重置/Del後再搜尋會閃退的bug 部分method/func 回傳型態修改 圖書館與新生週按鈕隱藏 手動merge cal_fix

+++ NTOU_Mobile/module/ShuttleTrack/NTOUTableViewControllerLayer1.m
@@ -68,7 +68,7 @@
 #pragma mark - Table view data source
 
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     CGFloat rowHeight = 0;
     UIFont *cellFont = [UIFont fontWithName:@"Helvetica" size:14.0];
