+++ NTOU_Mobile/module/ShuttleTrack/R66Layer2ViewController.m
@@ -88,7 +88,7 @@
     return @"例假日發車時間";
 }
 
--(float)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
+-(CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section
 {
     return 32;
 }
